==> CNN_source_code/NB4_v6_Training_CNN_NER_w266_FP-without-outputs.ipynb <==
import os
import json
import joblib
import time

import pandas as pd
import random
from IPython.display import display, HTML

import torch
import transformers
from transformers import BartTokenizer, BartForConditionalGeneration

#connect to hugging face
!git clone https://github.com/huggingface/transformers \
&& cd transformers \
install
!pip install -q ./transformers

# Installing transformers
!pip install simpletransformers
!pip install transformers
# Importing necessary libraries from simple transformers
import pandas as pd
from simpletransformers.seq2seq import Seq2SeqModel,Seq2SeqArgs

pip install ipywidgets

device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
print('Using: ', device)

headlines_processed = []
clean_news = []
dataset = []
target = []
count = 0

path_models = './models/'
path_outputs = './outputs/'
path_processed_data= './processed_data/'

jsonFile = open(os.path.join(path_processed_data, 'cnn_train_v3.json'), "r")
jsonContent = jsonFile.read()
details_dict = json.loads(jsonContent)
jsonFile.close()

X_train = details_dict['text_ner']
Y_train = details_dict['headlines']

jsonFile = open(os.path.join(path_processed_data, 'cnn_valid_v3.json'), "r")
jsonContent = jsonFile.read()
details_dict = json.loads(jsonContent)
jsonFile.close()

X_valid = details_dict['text_ner']
Y_valid = details_dict['headlines']

jsonFile = open(os.path.join(path_processed_data, 'cnn_test_v3.json'), "r")
jsonContent = jsonFile.read()
details_dict = json.loads(jsonContent)
jsonFile.close()

X_test = details_dict['text_ner']
Y_test = details_dict['headlines']


print(len(X_train), len(Y_train), len(X_valid), len(Y_valid), len(X_test), len(Y_test))

X_train[1:2]

Y_train[1:2]

details = {
    'headlines' : Y_train,
    'text' : X_train
}

df_train = pd.DataFrame(details)

details = {
    'headlines' : Y_valid,
    'text' : X_valid
}

df_valid = pd.DataFrame(details)

details = {
    'headlines' : Y_test,
    'text' : X_test
}

df_test = pd.DataFrame(details)

# Import


tokenizer = BartTokenizer.from_pretrained("facebook/bart-large")
model = BartForConditionalGeneration.from_pretrained("facebook/bart-large")

torch_device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
print('Using: ', torch_device)


# Rename columns as per pretrained model required format
df_train=df_train.rename(columns={'headlines':'target_text','text':'input_text'})
df_valid=df_valid.rename(columns={'headlines':'target_text','text':'input_text'})
df_test=df_test.rename(columns={'headlines':'target_text','text':'input_text'})
#df = {dataset, target}
model_args = Seq2SeqArgs()
# Initializing number of epochs
#model_args.num_train_epochs = 25
model_args.num_train_epochs = 2
# Initializing no_save arg
model_args.no_save = True
model_args.overwrite_output_dir = True

# Initializing evaluation args
model_args.evaluate_generated_text = True
model_args.evaluate_during_training = True
model_args.evaluate_during_training_verbose = True

df_train['target_text'][0]

df_train['input_text'][0]

df_valid['target_text'][0]

df_valid['input_text'][0]

df_test['target_text'][0]

df_test['input_text'][0]

model = Seq2SeqModel(
encoder_decoder_type="bart",
encoder_decoder_name="facebook/bart-large",
args=model_args,
use_cuda=True,
)
# Splitting data into train-test
# from sklearn.model_selection import train_test_split
# train_df, test_df = train_test_split(df, test_size=0.2)


df_train.shape, df_valid.shape, df_test.shape

len(df_train)//10000

def current_time_min():
    time_in_mins = time.time()/60.0
    return round(time_in_mins, 4)

# Training the model and keeping eval dataset as test data
# 16000 for aws, 800 for local VM
batch_size_train = 10000
batch_size_test = batch_size_train//4
number_of_batches = len(df_train)//batch_size_train
count = 0
epocs = model_args.num_train_epochs
acc_time = 0

for i in range(number_of_batches):
    start_time = current_time_min()
    count += 1
    start_offset_train = i * batch_size_train
    start_offset_test = i * batch_size_test
    end_offset_train = start_offset_train + batch_size_train
    end_offset_test = start_offset_test + batch_size_test
    model.train_model(df_train[start_offset_train:end_offset_train], eval_data=df_valid)
    model_filename = "model_ner_" + str(count) + "_" + str(start_offset_train) + "_" + str(end_offset_train) + "_" + str(epocs) + "e.pkl"
    joblib.dump(model, os.path.join(path_models, model_filename))
    results = model.eval_model(df_test)
    batch_time = current_time_min() - start_time
    acc_time = acc_time + batch_time/60
    randint = random.randrange(1, len(df_test), 3)
    results_json = {
        'iteration' : count,
        'start' : start_offset_train,
        'end' : end_offset_train,
        'eval_loss' : results["eval_loss"],
        'batch_time' : batch_time,
        'acc_time' : acc_time
    }
    jsonString = json.dumps(results_json)
    results_filename = "results_" + str(start_offset_train) + "_" + str(end_offset_train) + "_" + str(count)
    results_file = open(os.path.join(path_models, "results.txt"), "a")
    results_file.write(jsonString + '\n')
    results_file.close()
    print("Count: ", count, "Results: ", results)


jsonFile = open(os.path.join(path_processed_data, 'cnn_train_v3.json'), "r")
jsonContent = jsonFile.read()
details_dict = json.loads(jsonContent)
jsonFile.close()

X_train = details_dict['text']
Y_train = details_dict['headlines']

jsonFile = open(os.path.join(path_processed_data, 'cnn_valid_v3.json'), "r")
jsonContent = jsonFile.read()
details_dict = json.loads(jsonContent)
jsonFile.close()

X_valid = details_dict['text']
Y_valid = details_dict['headlines']

jsonFile = open(os.path.join(path_processed_data, 'cnn_test_v3.json'), "r")
jsonContent = jsonFile.read()
details_dict = json.loads(jsonContent)
jsonFile.close()

X_test = details_dict['text']
Y_test = details_dict['headlines']


details = {
    'headlines' : Y_train,
    'text' : X_train
}

df_train = pd.DataFrame(details)

details = {
    'headlines' : Y_valid,
    'text' : X_valid
}

df_valid = pd.DataFrame(details)

details = {
    'headlines' : Y_test,
    'text' : X_test
}

df_test = pd.DataFrame(details)



# Import


tokenizer = BartTokenizer.from_pretrained("facebook/bart-large")
model = BartForConditionalGeneration.from_pretrained("facebook/bart-large")

torch_device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
print('Using: ', torch_device)

# Rename columns as per pretrained model required format
df_train=df_train.rename(columns={'headlines':'target_text','text':'input_text'})
df_valid=df_valid.rename(columns={'headlines':'target_text','text':'input_text'})
df_test=df_test.rename(columns={'headlines':'target_text','text':'input_text'})
#df = {dataset, target}
model_args = Seq2SeqArgs()
# Initializing number of epochs
#model_args.num_train_epochs = 25
model_args.num_train_epochs = 2
# Initializing no_save arg
model_args.no_save = True
model_args.overwrite_output_dir = True

# Initializing evaluation args
model_args.evaluate_generated_text = True
model_args.evaluate_during_training = True
model_args.evaluate_during_training_verbose = True

model = Seq2SeqModel(
encoder_decoder_type="bart",
encoder_decoder_name="facebook/bart-large",
args=model_args,
use_cuda=True,
)

# Training the model and keeping eval dataset as test data
# 16000 for aws, 800 for local VM
batch_size_train = 10000
batch_size_test = batch_size_train//4
number_of_batches = len(df_train)//batch_size_train
count = 0
epocs = model_args.num_train_epochs
acc_time = 0

for i in range(number_of_batches):
    start_time = current_time_min()
    count += 1
    start_offset_train = i * batch_size_train
    start_offset_test = i * batch_size_test
    end_offset_train = start_offset_train + batch_size_train
    end_offset_test = start_offset_test + batch_size_test
    model.train_model(df_train[start_offset_train:end_offset_train], eval_data=df_valid)
    model_filename = "model_" + str(count) + "_" + str(start_offset_train) + "_" + str(end_offset_train) + "_" + str(epocs) + "e.pkl"
    joblib.dump(model, os.path.join(path_models, model_filename))
    results = model.eval_model(df_test)
    batch_time = current_time_min() - start_time
    acc_time = acc_time + batch_time/60
    randint = random.randrange(1, len(df_test), 3)
    results_json = {
        'iteration' : count,
        'start' : start_offset_train,
        'end' : end_offset_train,
        'eval_loss' : results["eval_loss"],
        'batch_time' : batch_time,
        'acc_time' : acc_time
    }
    jsonString = json.dumps(results_json)
    results_filename = "results_" + str(start_offset_train) + "_" + str(end_offset_train) + "_" + str(count)
    results_file = open(os.path.join(path_models, "results_2.txt"), "a")
    results_file.write(jsonString + '\n')
    results_file.close()
    print("Count: ", count, "Results: ", results)
