==> CNN_source_code/NB4_v6_Training_CNN_T5_w266_FP.ipynb <==
import os
import json
import joblib
import time

import pandas as pd
import random
from IPython.display import display, HTML
import torch
import numpy as np

# import torch
# import transformers
# from transformers import BartTokenizer, BartForConditionalGeneration

# #connect to hugging face
# !git clone https://github.com/huggingface/transformers \
# && cd transformers \
# install
#!pip install -q ./transformers

!pip install transformers==4.16.2

# # Installing transformers
# !pip install simpletransformers
# !pip install transformers
# # Importing necessary libraries from simple transformers
# import pandas as pd
# from simpletransformers.seq2seq import Seq2SeqModel,Seq2SeqArgs

# pip install ipywidgets

# device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
# print('Using: ', device)

# headlines_processed = []
# clean_news = []
# dataset = []
# target = []
# count = 0

path_models = './models/'
path_outputs = './outputs/'
path_processed_data= './processed_data/'

jsonFile = open(os.path.join(path_processed_data, 'cnn_train_v3.json'), "r")
jsonContent = jsonFile.read()
details_dict = json.loads(jsonContent)
jsonFile.close()

X_train = details_dict['text']
Y_train = details_dict['headlines']

jsonFile = open(os.path.join(path_processed_data, 'cnn_valid_v3.json'), "r")
jsonContent = jsonFile.read()
details_dict = json.loads(jsonContent)
jsonFile.close()

X_valid = details_dict['text']
Y_valid = details_dict['headlines']

jsonFile = open(os.path.join(path_processed_data, 'cnn_test_v3.json'), "r")
jsonContent = jsonFile.read()
details_dict = json.loads(jsonContent)
jsonFile.close()

X_test = details_dict['text']
Y_test = details_dict['headlines']


print(len(X_train), len(Y_train), len(X_valid), len(Y_valid), len(X_test), len(Y_test))

details = {
    'headlines' : Y_train,
    'text' : X_train
}

df_train = pd.DataFrame(details)

details = {
    'headlines' : Y_valid,
    'text' : X_valid
}

df_valid = pd.DataFrame(details)

details = {
    'headlines' : Y_test,
    'text' : X_test
}

df_test = pd.DataFrame(details)

# # Import
# import numpy as np # linear algebra
# import pandas as pd
# import logging
# import os

# logging.basicConfig(level=logging.INFO)
# transformers_logger = logging.getLogger("transformers")
# transformers_logger.setLevel(logging.WARNING)

# torch_device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
# print('Using: ', torch_device)


import os, psutil  

def cpu_stats():
    pid = os.getpid()
    py = psutil.Process(pid)
    memory_use = py.memory_info()[0] / 2. ** 30
    return 'memory GB:' + str(np.round(memory_use, 2))

def current_time_min():
    time_in_mins = time.time()/60.0
    return round(time_in_mins, 4)

cpu_stats()

# import logging

# import pandas as pd
# from simpletransformers.t5 import T5Model

# logging.basicConfig(level=logging.INFO)
# transformers_logger = logging.getLogger("transformers")
# transformers_logger.setLevel(logging.WARNING)

df_train['prefix'] = "summarize"
df_test['prefix'] = "summarize"
df_valid['prefix'] = "summarize"

model_args = {
    "reprocess_input_data": True,
    "overwrite_output_dir": True,
    "max_seq_length": 512,
    "train_batch_size": 16,
    "num_train_epochs": 10,
}



# Create T5 Model
#model = T5Model(model_type="t5", model_name="t5-base", args=model_args, use_cuda=True)

# Rename columns as per pretrained model required format
df_train=df_train.rename(columns={'headlines':'target_text','text':'source_text'})
df_valid=df_valid.rename(columns={'headlines':'target_text','text':'source_text'})
df_test=df_test.rename(columns={'headlines':'target_text','text':'source_text'})


df_train.shape, df_valid.shape, df_test.shape

# len(df_train)//10000

# start_time = current_time_min()
# model.train_model(df_train[:100])
# batch_time = current_time_min() - start_time
# print(batch_time)


!pip install simplet5

from simplet5 import SimpleT5


model = SimpleT5()
#model.from_pretrained(model_type="t5", model_name="t5-base")
model.from_pretrained(model_type="t5", model_name="t5-large")

model.train(train_df=df_train,
            eval_df=df_valid, 
            source_max_token_len=128, 
            target_max_token_len=50, 
            batch_size=16, 
            max_epochs=10, 
            use_gpu=True
           )

df_test['source_text'] = "summarize: " + df_test['source_text']

df_test.head()

# model_filename = "model_t5_3e.pkl"
# joblib.dump(model, os.path.join(path_models, model_filename))

#model = joblib.load(os.path.join(path_models, 'model_t5_3e.pkl'))
model.load_model("t5", "outputs/simplet5-epoch-5-train-loss-2.2146-val-loss-2.8844", use_gpu=True)

t5_prepared_Text = "summarize: " + df_test[0:1]['source_text'][0]
P1 = model.predict(t5_prepared_Text)

P1

#from transformers import T5Tokenizer, T5ForConditionalGeneration, T5Config

#model = T5ForConditionalGeneration.from_pretrained('t5-small')
#tokenizer = T5Tokenizer.from_pretrained(model_type="t5", model_name="t5-base", pretrained_model_name_or_path='./models/model_t5_3e.pkl')

t5_prepared_Text = "summarize: " + df_test[0:1]['source_text'][0]
#tokenized_text = tokenizer.encode(t5_prepared_Text, max_length=2048, return_tensors="pt", truncation=True).to(device) #default 512
        
summary_ids = model.predict(t5_prepared_Text,
                            num_beams=4,
                            max_length=60,
                            early_stopping=True,
                            use_gpu=True)

summary_t5 = tokenizer.decode(summary_ids[0], skip_special_tokens=True)

# P1 = model.predict(df_test[0:1]['source_text'][0])
# #df_test[0:1]['source_text'][0]

# Training the model and keeping eval dataset as test data
# 16000 for aws, 800 for local VM
#batch_size_train = 10000
batch_size_train = 100

batch_size_test = batch_size_train//4
number_of_batches = len(df_train)//batch_size_train
#count = 0
epocs = model_args['num_train_epochs']
#acc_time = 0

for i in range(number_of_batches):
    start_time = current_time_min()
    count += 1
    start_offset_train = i * batch_size_train
    start_offset_test = i * batch_size_test
    end_offset_train = start_offset_train + batch_size_train
    end_offset_test = start_offset_test + batch_size_test
    model.train_model(df_train[start_offset_train:end_offset_train], eval_data=df_valid)
    model_filename = "model_" + str(count) + "_" + str(start_offset_train) + "_" + str(end_offset_train) + "_" + str(epocs) + "e.pkl"
    joblib.dump(model, os.path.join(path_models, model_filename))
    results = model.eval_model(df_test)
    batch_time = current_time_min() - start_time
    acc_time = acc_time + batch_time/60
    randint = random.randrange(1, len(df_test), 3)
    results_json = {
        'iteration' : count,
        'start' : start_offset_train,
        'end' : end_offset_train,
        'eval_loss' : results["eval_loss"],
        'batch_time' : batch_time,
        'acc_time' : acc_time
    }
    jsonString = json.dumps(results_json)
    results_filename = "results_" + str(start_offset_train) + "_" + str(end_offset_train) + "_" + str(count)
    results_file = open(os.path.join(path_models, "results.txt"), "a")
    results_file.write(jsonString + '\n')
    results_file.close()
    print("Count: ", count, "Results: ", results)


# Generating summaries on news test data
results = model.eval_model(df_test[100:200])
# print the loss
results



# MODEL_PATH = os.path.join(path_models, 'bart.model')

# torch.save(model, MODEL_PATH)

joblib.dump(model, os.path.join(path_models, 'model.pkl'))

model = Seq2SeqModel(encoder_decoder_type="bart",
  encoder_decoder_name="facebook/bart-large",
  args=model_args,
  use_cuda=False,)

model.load(torch.load(MODEL_PATH))


model = joblib.load(os.path.join(path_models, 'model.pkl'))

# model.eval()

# Original test data text summary for top 10 news
#value = randint(0, 10)
for i in test_df.target_text[422:423]:
  print(i)



# Predicted summary
for i in test_df.input_text[1000:1050]:
  print(model.predict([i]))
