==> CNN_source_code/NB6_v3_Inference_CNN_T5_w266_FP.ipynb <==
import os
import json
import joblib
import time
import pandas as pd
import numpy as np
import random
from IPython.display import display, HTML
import torch

# Better pandas table
# from google.colab import data_table
# data_table.enable_dataframe_formatter()

#connect to hugging face
!git clone https://github.com/huggingface/transformers \
&& cd transformers \
install
!pip install -q ./transformers

import transformers
from transformers import BartTokenizer, BartForConditionalGeneration

# Installing transformers
!pip install simpletransformers
#!pip install transformers
# Importing necessary libraries from simple transformers
from simpletransformers.seq2seq import Seq2SeqModel,Seq2SeqArgs

pip install ipywidgets

device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
print('Using: ', device)

# from google.colab import drive
# drive.mount('/content/drive')

# path_models = '/content/drive/MyDrive/w266/models/'
# path_outputs = '/content/drive/MyDrive/w266/outputs/'
# path_processed_data= '/content/drive/MyDrive/w266/processed_data/'

path_models = './models/'
path_outputs = './outputs/'
path_processed_data= './processed_data/'

data_name = 'cnn_test'
input_filename = data_name + '_summary_v3.json' 
output_filename = data_name + '_prediction_t5_v3.json'

jsonFile = open(os.path.join(path_processed_data, input_filename), "r")
jsonContent = jsonFile.read()
details_dict = json.loads(jsonContent)
jsonFile.close()

details = {
    'headlines' : details_dict['headlines'],
    'text' : details_dict['text'],
    'keywords' : details_dict['keywords'],
    'summary_art' : details_dict['summary_art'],
    'entities' : details_dict['entities'],
#    'text_ner' : details_dict['text_ner'],
    'sentence_1s' : details_dict['entities'],
    'sentence_3s' : details_dict['sentence_3s'],
    'summary_ext' : details_dict['summary_ext'],
    'summary_abs' : details_dict['summary_abs'],
    'summary_extabs' : details_dict['summary_extabs'],
    'summary_ner' : details_dict['summary_ner'],
    'summary_t5' : details_dict['summary_t5'],
    'predict_1s' : details_dict['predict_1s'],
    'predict_3s' : details_dict['predict_3s'],
    'predict_text' : details_dict['predict_text'],
    'predict_ext' : details_dict['predict_ext'],
    'predict_abs' : details_dict['predict_abs'],
    'predict_extabs' : details_dict['predict_extabs'],
    'predict_art' : details_dict['summary_art']
#     'predict_art' : details_dict['summary_art'],
#         'predict_art' : details_dict['summary_art']
}


df_test = pd.DataFrame(details)

print(len(details_dict['text_ner']))

# model = joblib.load(os.path.join(path_models, 'model_7_60000_70000_2e.pkl'))

import torch
import json 
from transformers import T5Tokenizer, T5ForConditionalGeneration, T5Config

model = T5ForConditionalGeneration.from_pretrained('t5-small')
tokenizer = T5Tokenizer.from_pretrained('t5-small')
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
device = torch.device("cpu")
print(device)

df = df_test[:10]

def summarize_text_t5(row):
    
    text = row[1]['text']
    #.split(' ', 512)
    #text = ' '.join([str(item) for item in text])
    max_len = min(len(text), 60)
    max_sentence_count = 5
    summary_text = []
    
    try:
        t5_prepared_Text = "summarize: " + text
        tokenized_text = tokenizer.encode(t5_prepared_Text, max_length=2048, truncation=True, return_tensors="pt").to(device) #default 512
        summary_ids = model.generate(tokenized_text,
                                    num_beams=4,
                                    no_repeat_ngram_size=2,
                                    min_length=30,
                                    max_length=max_len,
                                    early_stopping=True)

        summary_text = tokenizer.decode(summary_ids[0], skip_special_tokens=True)
        #summary_text = tokenizer.decode(summary_ids[0], skip_special_tokens=True)
        
    except Exception as error:
        print("Exception:", error)
        t5_prepared_Text = "summarize: " + text
        tokenized_text = tokenizer.encode(t5_prepared_Text, max_length=2048, return_tensors="pt").to(device) #default 512
        summary_ids = model.generate(tokenized_text,
                                    num_beams=4,
                                    no_repeat_ngram_size=2,
                                    min_length=30,
                                    max_length=max_len,
                                    early_stopping=True)

        summary_text = tokenizer.decode(summary_ids[0], skip_special_tokens=True)
        return text, summary_text
    return text, summary_text

# text = []
# headlines = []
# keywords = []
# entities = []
# sentence_1s = []
# sentence_3s = []
# summary_art = []
# summary_ext = []
# summary_abs = []
# summary_extabs = []
# summary_ner = []
summary_text = []

from joblib import Parallel, delayed
start = time.time()

df = df.reset_index()

S = Parallel(n_jobs=4)(delayed(summarize_text_t5)(row) for row in df.iterrows())

print("Response time (mins): ", round((time.time() - start)/60, 2))

summary_text = []
text = []
S = list(filter(None, S))
for i in range(len(S)):  
        text.append(S[i][0])  # text
        summary_text.append(S[i][1])  # summary_t5

summary_text

# articles = pd.DataFrame({
#     'headline': titles,
#     'text': abstracts
# })
# articles.head()


# prediction

# Prediction based on text
#P0 = model.predict(details_dict['sentence_1s'])




P1 = model.predict(details_dict['sentence_3s'])

# 'predict_1s' : P0,
# 'predict_3s' : P1,
# 'predict_text' : P2,
# 'predict_ext' : P3,
# 'predict_abs' : P4,
# 'predict_extabs' : P5,
# 'predict_ner' : P6
# 'summary_ext' : S2,
# 'summary_abs' : S3,
# 'summary_extabs' : S4,
# 'summary_ner' : S5,

count = 0
S2 = []
for item in details_dict['text']:
    if len(item) == 0:
        count += 1
        S2.append("None")
        #print(X[item])
    else:
        S2.append(item)

print(count)


P2 = model.predict(details_dict['text'])

P2

# 'predict_1s' : P0,
# 'predict_3s' : P1,
# 'predict_text' : P2,
# 'predict_ext' : P3,
# 'predict_abs' : P4,
# 'predict_extabs' : P5,
# 'predict_ner' : P6
# 'summary_ext' : S2,
# 'summary_abs' : S3,
# 'summary_extabs' : S4,
# 'summary_ner' : S5,

count = 0
S3 = []
for item in details_dict['summary_ext']:
    if len(item) == 0:
        count += 1
        S3.append("")
    else:
        S3.append(item)

print(count)

# Prediction based on extractive summary
P3 = model.predict(S3)



P3

count = 0
S4 = []
for item in details_dict['summary_abs']:
    if len(item) == 0:
        count += 1
        S4.append("")
        #print(X[item])
    else:
        S4.append(item)

print(count)

# Prediction based on abstractive summary
P4 = model.predict(S4)



P4

count = 0
S5 = []
for item in details_dict['summary_extabs']:
    if len(item) == 0:
        count += 1
        S5.append("")
        #print(X[item])
    else:
        S5.append(item)

print(count)

# Prediction based on ext-abs summary
P5 = model.predict(S5)



P5

count = 0
S6 = []
for item in details_dict['summary_ner']:
    if len(item) == 0:
        count += 1
        S6.append("None")
        #print(X[item])
    else:
        S6.append(item)

print(count)

# Prediction based on ext-abs-ner summary
P6 = model.predict(S6)

count = 0
S7 = []
for item in details_dict['summary_t5']:
    if len(item) == 0:
        count += 1
        S7.append("None")
        #print(X[item])
    else:
        S7.append(item)

print(count)

# Prediction based on t5 summary
P7 = model.predict(S7)

# Prediction based on art summary
P8 = model.predict(details_dict['summary_art'])


# details = {
#     'headlines' : details_dict['headlines'],
#     'text' : details_dict['text'],
#     'keywords' : details_dict['keywords'],
#     'summary_art' : details_dict['summary_art'],
#     'entities' : details_dict['entities'],
#     'sentence_1s' : details_dict['entities'],
#     'sentence_3s' : details_dict['sentence_3s'],
#     'summary_ext' : details_dict['summary_ext'],
#     'summary_abs' : details_dict['summary_abs'],
#     'summary_extabs' : details_dict['summary_extabs'],
#     'summary_ner' : details_dict['summary_ner'],
#     'summary_t5' : details_dict['summary_t5'],
#     'predict_1s' : details_dict['predict_t5'],
#     'predict_3s' : details_dict['predict_t5'],
#     'predict_text' : details_dict['predict_t5'],
#     'predict_ext' : details_dict['predict_t5'],
#     'predict_abs' : details_dict['predict_t5'],
#     'predict_extabs' : details_dict['predict_t5'],
#     'predict_ner' : P6,
#     'predict_t5' : details_dict['predict_t5'],
#     'predict_art' : details_dict['predict_t5'],
# }

details = {
    'headlines' : details_dict['headlines'],
    'text' : details_dict['text'],
    'keywords' : details_dict['keywords'],
    'summary_art' : details_dict['summary_art'],
    'entities' : details_dict['entities'],
    'sentence_1s' : details_dict['sentence_1s'],
    'sentence_3s' : details_dict['sentence_3s'],
    'summary_ext' : details_dict['summary_ext'],
    'summary_abs' : details_dict['summary_abs'],
    'summary_extabs' : details_dict['summary_extabs'],
    'summary_ner' : details_dict['summary_ner'],
    'summary_t5' : details_dict['summary_t5'],
    'predict_1s' : details_dict['sentence_1s'],
    'predict_3s' : P1,
    'predict_text' : P2,
    'predict_ext' : P3,
    'predict_abs' : P4,
    'predict_extabs' : P5,
#    'predict_ner' : P6,
#    'predict_t5' : details_dict['predict_t5'],
    'predict_art' : P8,
}


jsonString = json.dumps(details)
jsonFile = open(os.path.join(path_processed_data, output_filename), "w")
jsonFile.write(jsonString)
jsonFile.close()

df_results = {k: details[k] for k in ('headlines', 'predict_3s', 'predict_text', 'predict_ext', 'predict_abs', 'predict_extabs', 'predict_art')}
pd.DataFrame(df_results)
