==> CNN_source_code/NB7_v1_Scoring_CNN_w266_FP.ipynb <==
import os
import json
import joblib
import time
import pandas as pd
import numpy as np
import random
from IPython.display import display, HTML
import torch

import nltk

# # Better pandas table
# from google.colab import data_table
# data_table.enable_dataframe_formatter()

#connect to hugging face
!git clone https://github.com/huggingface/transformers \
&& cd transformers \
install
!pip install -q ./transformers

import transformers
from transformers import BartTokenizer, BartForConditionalGeneration

# Installing transformers
!pip install simpletransformers
#!pip install transformers
# Importing necessary libraries from simple transformers
from simpletransformers.seq2seq import Seq2SeqModel,Seq2SeqArgs

pip install ipywidgets

device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
print('Using: ', device)

# from google.colab import drive
# drive.mount('/content/drive')

# path_models = '/content/drive/MyDrive/w266/models/'
# path_outputs = '/content/drive/MyDrive/w266/outputs/'
# path_processed_data= '/content/drive/MyDrive/w266/processed_data/'

path_models = './models/cnn_2e/'
path_outputs = './outputs/'
path_processed_data= './processed_data/'

data_name = 'cnn_test'
input_filename = data_name + '_prediction_v3_bart_cnn.json' 
#out_filename = data_name + '_summary_abs.json'


jsonFile = open(os.path.join(path_processed_data, input_filename), "r")
jsonContent = jsonFile.read()
details_dict = json.loads(jsonContent)
jsonFile.close()

details = {
    'headlines' : details_dict['headlines'],
    'text' : details_dict['text'],
    'keywords' : details_dict['keywords'],
    'summary_art' : details_dict['summary_art'],
    'entities' : details_dict['entities'],
#    'text_ner' : details_dict['text_ner'],
    'sentence_1s' : details_dict['sentence_1s'],
    'sentence_3s' : details_dict['sentence_3s'],
    'summary_ext' : details_dict['summary_ext'],
    'summary_abs' : details_dict['summary_abs'],
    'summary_extabs' : details_dict['summary_extabs'],
    'summary_ner' : details_dict['summary_ner'],
    'summary_t5' : details_dict['summary_t5'],
    'predict_1s' : details_dict['predict_1s'],
    'predict_3s' : details_dict['predict_3s'],
    'predict_text' : details_dict['predict_text'],
    'predict_ext' : details_dict['predict_ext'],
    'predict_abs' : details_dict['predict_abs'],
    'predict_extabs' : details_dict['predict_extabs'],
    'predict_art' : details_dict['predict_art'],
#    'predict_t5' : details_dict['predict_t5']
}

df_score = pd.DataFrame(details)

#print(len(details_dict['predict_ner']))

df_results = {k: details[k] for k in ('headlines', 'predict_text', 'predict_ext', 'predict_art', 'predict_3s', 'predict_abs', 'predict_extabs', 'predict_1s')}
df = pd.DataFrame(df_results)

pd.set_option("display.max_rows", None, "display.max_columns", None)

import itertools
n = 9
new_dict = {}
first_n_pairs = itertools.islice(df_results.items(), n)
for key, value in first_n_pairs:
    #print(key, '::', value[3])
    new_dict[key] = value[424]

new_df = pd.DataFrame(new_dict.items(), columns=['Approach', 'Headlines'])
#new_df.style.set_properties(**{'text-align': 'left'})
new_df.style.set_properties(**{'text-align': 'left'}).set_table_styles([ dict(selector='th', props=[('text-align', 'left')] ) ])
new_df

display(HTML(new_df.to_html(justify='left').replace('<tr>','<tr style="text-align: left;">')))

df_results = {k: details[k] for k in ('text', 'sentence_1s', 'sentence_3s', 'summary_ext', 'summary_abs', 'summary_extabs', 'summary_art')}
df = pd.DataFrame(df_results)

pd.set_option("display.max_rows", None, "display.max_columns", None)

import itertools
n = 7
new_dict = {}
first_n_pairs = itertools.islice(df_results.items(), n)
for key, value in first_n_pairs:
    #print(key, '::', value[3])
    new_dict[key] = value[424]

new_df = pd.DataFrame(new_dict.items(), columns=['Summary', 'Content'])
#new_df.style.set_properties(**{'text-align': 'left'})
new_df.style.set_properties(**{'text-align': 'left'}).set_table_styles([ dict(selector='th', props=[('text-align', 'left')] ) ])
new_df

pip install Rouge

from rouge import Rouge
rouge = Rouge()

score_1s = rouge.get_scores(details_dict['predict_1s'], details_dict['headlines'], avg=True)
df_results = {k: score_1s[k] for k in ('rouge-1', 'rouge-2', 'rouge-l')}
pd.DataFrame(df_results)

score_3s = rouge.get_scores(details_dict['predict_3s'], details_dict['headlines'], avg=True)
df_results = {k: score_3s[k] for k in ('rouge-1', 'rouge-2', 'rouge-l')}
pd.DataFrame(df_results)

score_text = rouge.get_scores(details_dict['predict_text'], details_dict['headlines'], avg=True)
df_results = {k: score_text[k] for k in ('rouge-1', 'rouge-2', 'rouge-l')}
pd.DataFrame(df_results)

score_ext = rouge.get_scores(details_dict['predict_ext'], details_dict['headlines'], avg=True)
df_results = {k: score_ext[k] for k in ('rouge-1', 'rouge-2', 'rouge-l')}
pd.DataFrame(df_results)

score_abs = rouge.get_scores(details_dict['predict_abs'], details_dict['headlines'], avg=True)
df_results = {k: score_abs[k] for k in ('rouge-1', 'rouge-2', 'rouge-l')}
pd.DataFrame(df_results)

score_extabs = rouge.get_scores(details_dict['predict_extabs'], details_dict['headlines'], avg=True)
df_results = {k: score_extabs[k] for k in ('rouge-1', 'rouge-2', 'rouge-l')}
pd.DataFrame(df_results)

# score_ner = rouge.get_scores(details_dict['predict_ner'], details_dict['headlines'], avg=True)
# df_results = {k: score_ner[k] for k in ('rouge-1', 'rouge-2', 'rouge-l')}
# pd.DataFrame(df_results)

# score_t5 = rouge.get_scores(details_dict['predict_t5'], details_dict['headlines'], avg=True)
# df_results = {k: score_t5[k] for k in ('rouge-1', 'rouge-2', 'rouge-l')}
# pd.DataFrame(df_results)

score_art = rouge.get_scores(details_dict['predict_art'], details_dict['headlines'], avg=True)
df_results = {k: score_art[k] for k in ('rouge-1', 'rouge-2', 'rouge-l')}
pd.DataFrame(df_results)

rouge_1 = {'baseline (1s)': {'r': score_1s['rouge-1']['r'], 'p':score_1s['rouge-1']['p'], 'f': score_1s['rouge-1']['f']}, 
        'baseline (3s)': {'r': score_3s['rouge-1']['r'], 'p':score_3s['rouge-1']['p'], 'f': score_3s['rouge-1']['f']}, 
           'text': {'r': score_text['rouge-1']['r'], 'p':score_text['rouge-1']['p'], 'f': score_text['rouge-1']['f']},
           'ext': {'r': score_ext['rouge-1']['r'], 'p':score_ext['rouge-1']['p'], 'f': score_ext['rouge-1']['f']},
           'abs': {'r': score_abs['rouge-1']['r'], 'p':score_abs['rouge-1']['p'], 'f': score_abs['rouge-1']['f']},
           'extabs': {'r': score_extabs['rouge-1']['r'], 'p':score_extabs['rouge-1']['p'], 'f': score_extabs['rouge-1']['f']},
           'artsum': {'r': score_art['rouge-1']['r'], 'p':score_art['rouge-1']['p'], 'f': score_art['rouge-1']['f']},
#           't5': {'r': score_t5['rouge-1']['r'], 'p':score_t5['rouge-1']['p'], 'f': score_t5['rouge-1']['f']},
}

df_results = {k: rouge_1[k] for k in ('artsum', 'text', 'ext', 'baseline (3s)',  'baseline (1s)', 'abs', 'extabs', )}
pd.DataFrame(df_results)

rouge_2 = {'baseline (1s)': {'r': score_1s['rouge-2']['r'], 'p':score_1s['rouge-2']['p'], 'f': score_1s['rouge-2']['f']}, 
        'baseline (3s)': {'r': score_3s['rouge-2']['r'], 'p':score_3s['rouge-2']['p'], 'f': score_3s['rouge-2']['f']}, 
           'text': {'r': score_text['rouge-2']['r'], 'p':score_text['rouge-2']['p'], 'f': score_text['rouge-2']['f']},
           'ext': {'r': score_ext['rouge-2']['r'], 'p':score_ext['rouge-2']['p'], 'f': score_ext['rouge-2']['f']},
           'abs': {'r': score_abs['rouge-2']['r'], 'p':score_abs['rouge-2']['p'], 'f': score_abs['rouge-2']['f']},
           'extabs': {'r': score_extabs['rouge-2']['r'], 'p':score_extabs['rouge-2']['p'], 'f': score_extabs['rouge-2']['f']},
           'artsum': {'r': score_art['rouge-2']['r'], 'p':score_art['rouge-2']['p'], 'f': score_art['rouge-2']['f']},
#           't5': {'r': score_t5['rouge-2']['r'], 'p':score_t5['rouge-2']['p'], 'f': score_t5['rouge-2']['f']},
}
df_results = {k: rouge_2[k] for k in ('artsum', 'text', 'ext', 'baseline (3s)',  'baseline (1s)', 'abs', 'extabs', )}
pd.DataFrame(df_results)

rouge_l = {'baseline (1s)': {'r': score_1s['rouge-l']['r'], 'p':score_1s['rouge-l']['p'], 'f': score_1s['rouge-l']['f']}, 
        'baseline (3s)': {'r': score_3s['rouge-l']['r'], 'p':score_3s['rouge-l']['p'], 'f': score_3s['rouge-l']['f']}, 
           'text': {'r': score_text['rouge-l']['r'], 'p':score_text['rouge-l']['p'], 'f': score_text['rouge-l']['f']},
           'ext': {'r': score_ext['rouge-l']['r'], 'p':score_ext['rouge-l']['p'], 'f': score_ext['rouge-l']['f']},
           'abs': {'r': score_abs['rouge-l']['r'], 'p':score_abs['rouge-l']['p'], 'f': score_abs['rouge-l']['f']},
           'extabs': {'r': score_extabs['rouge-l']['r'], 'p':score_extabs['rouge-l']['p'], 'f': score_extabs['rouge-l']['f']},
           'artsum': {'r': score_art['rouge-l']['r'], 'p':score_art['rouge-l']['p'], 'f': score_art['rouge-l']['f']},
#           't5': {'r': score_t5['rouge-l']['r'], 'p':score_t5['rouge-l']['p'], 'f': score_t5['rouge-l']['f']},
}
df_results = {k: rouge_l[k] for k in ('artsum', 'text', 'ext', 'baseline (3s)',  'baseline (1s)', 'abs', 'extabs', )}
pd.DataFrame(df_results)

# Evaluation Metrics - BLEU implementation
from nltk.translate.bleu_score import sentence_bleu


# Function to calculate the score
def calc_bleu(x, y):
    L = 0
    for i, j in zip(x, y):
      L += sentence_bleu([i], j, weights=(0.25, 0.25, 0.25, 0.25), smoothing_function=None,  auto_reweigh=False,)

    # Average BLEU score of whole corpuses
    return round(L/len(x), 4)



bleu_1s = calc_bleu(details_dict['headlines'], details_dict['predict_1s'])
print("BLEU score: {:.04f}".format(bleu_1s))

bleu_3s = calc_bleu(details_dict['headlines'], details_dict['predict_3s'])
print("BLEU score: {:.04f}".format(bleu_3s))

bleu_text = calc_bleu(details_dict['headlines'], details_dict['predict_text'])
print("BLEU score: {:.04f}".format(bleu_text))

bleu_ext = calc_bleu(details_dict['headlines'], details_dict['predict_ext'])
print("BLEU score: {:.04f}".format(bleu_ext))

bleu_abs = calc_bleu(details_dict['headlines'], details_dict['predict_abs'])
print("BLEU score: {:.04f}".format(bleu_abs))

bleu_extabs = calc_bleu(details_dict['headlines'], details_dict['predict_extabs'])
print("BLEU score: {:.04f}".format(bleu_extabs))

# bleu_ner = calc_bleu(details_dict['headlines'], details_dict['predict_ner'])
# print("BLEU score: {:.04f}".format(bleu_ner))

# bleu_t5 = calc_bleu(details_dict['headlines'], details_dict['predict_t5'])
# print("BLEU score: {:.04f}".format(bleu_t5))

bleu_art = calc_bleu(details_dict['headlines'], details_dict['predict_art'])
print("BLEU score: {:.04f}".format(bleu_art))

bleu_score = {'baseline (1s)': {'bleu_score': bleu_1s}, 
                'baseline (3s)': {'bleu_score': bleu_3s}, 
    'text' : {'bleu_score': bleu_text}, 
    'ext' : {'bleu_score': bleu_ext},
    'abs' : {'bleu_score': bleu_abs},
    'extabs' : {'bleu_score': bleu_extabs},
    'artsum' : {'bleu_score': bleu_art},
#    't5' : {'bleu_score': bleu_t5}
}

df_results = {k: bleu_score[k] for k in ('artsum', 'text', 'ext', 'baseline (3s)',  'abs', 'extabs', 'baseline (1s)')}
pd.DataFrame(df_results)

pip install bert-score

from bert_score import score

# Baseline
P, R, F1 = score(details_dict['predict_1s'], details_dict['headlines'], lang="en", verbose=True)
F1_1s = F1.mean().item()
P_1s = P.mean().item()
R_1s = R.mean().item()

# Ext
P, R, F1 = score(details_dict['predict_3s'], details_dict['headlines'], lang="en", verbose=True)
F1_3s = F1.mean().item()
P_3s = P.mean().item()
R_3s = R.mean().item()


# Abs
P, R, F1 = score(details_dict['predict_text'], details_dict['headlines'], lang="en", verbose=True)
F1_text = F1.mean().item()
P_text = P.mean().item()
R_text = R.mean().item()

# ExtAbs
P, R, F1 = score(details_dict['predict_ext'], details_dict['headlines'], lang="en", verbose=True)
F1_ext = F1.mean().item()
P_ext = P.mean().item()
R_ext = R.mean().item()

P, R, F1 = score(details_dict['predict_abs'], details_dict['headlines'], lang="en", verbose=True)
F1_abs = F1.mean().item()
P_abs = P.mean().item()
R_abs = R.mean().item()

P, R, F1 = score(details_dict['predict_extabs'], details_dict['headlines'], lang="en", verbose=True)
F1_extabs = F1.mean().item()
P_extabs = P.mean().item()
R_extabs = R.mean().item()

# P, R, F1 = score(details_dict['predict_ner'], details_dict['headlines'], lang="en", verbose=True)
# F1_ner = F1.mean().item()
# P_ner = P.mean().item()
# R_ner = R.mean().item()

# P, R, F1 = score(details_dict['predict_t5'], details_dict['headlines'], lang="en", verbose=True)
# F1_t5 = F1.mean().item()
# P_t5 = P.mean().item()
# R_t5 = R.mean().item()

P, R, F1 = score(details_dict['predict_art'], details_dict['headlines'], lang="en", verbose=True)
F1_art = F1.mean().item()
P_art = P.mean().item()
R_art = R.mean().item()

# import matplotlib.pyplot as plt
# plt.hist(F1, bins=20)
# plt.show()
# from bert_score import plot_example

# p1 = P1[0]
# y = Y[0]
# plot_example(p1, y, lang="en")

# bert_score = {'baseline': {'P': P_base,
#   'R': R_base,
#   'F1': F1_base}, 
#    'ext' : {'P': P_ext,
#   'R': R_ext,
#   'F1': F1_ext}, 
#   'abs' : {'P': P_abs,
#   'R': R_abs,
#   'F1': F1_abs},
#   'extabs' : {'P': P_extabs,
#   'R': R_extabs,
#   'F1': F1_extabs},
# }

bert_score = {'P': {  'art': P_art,
                      'text': P_text,
                      'ext': P_ext,
  'baseline (3 sentences)': P_3s,

  'abs': P_abs,
  'extabs': P_extabs,
'baseline (1 sentence)': P_1s,}, 
   'R' : {  'art': R_art,
            'text': R_text,
            'ext': R_ext,
  'baseline (3 sentences)': R_3s,
  'abs': R_abs,
  'extabs': R_extabs,
   'baseline (1 sentence)': R_1s,}, 
  'F1' : {  'art': F1_art,
            'text': F1_text,
            'ext': F1_ext,
  'baseline (3 sentences)': F1_3s,
  'abs': F1_abs,
  'extabs': F1_extabs,
  'baseline (1 sentence)': F1_1s,},
}



df_results = {k: bert_score[k] for k in ('P', 'R', 'F1')}
pd.DataFrame(df_results)

#!pip install --upgrade pip  # ensures that pip is current
#!git clone https://github.com/google-research/bleurt.git
#!cd bleurt
#!pip install .

# import unittest

# python -m unittest bleurt.score_test
# python -m unittest bleurt.score_not_eager_test
# python -m unittest bleurt.finetune_test
# python -m unittest bleurt.score_files_test

from functools import reduce
  
def average(lst):
    return reduce(lambda a, b: a + b, lst) / len(lst)

from bleurt import score

checkpoint = "bleurt/bleurt/test_checkpoint"
scorer = score.BleurtScorer(checkpoint)

scores = scorer.score(references=details_dict['headlines'], candidates=details_dict['predict_1s']) 
bleurt_1s = average(scores)

scores = scorer.score(references=details_dict['headlines'], candidates=details_dict['predict_3s']) 
bleurt_3s = average(scores)

scores = scorer.score(references=details_dict['headlines'], candidates=details_dict['predict_text']) 
bleurt_text = average(scores)

scores = scorer.score(references=details_dict['headlines'], candidates=details_dict['predict_ext']) 
bleurt_ext = average(scores)

scores = scorer.score(references=details_dict['headlines'], candidates=details_dict['predict_abs']) 
bleurt_abs = average(scores)

scores = scorer.score(references=details_dict['headlines'], candidates=details_dict['predict_extabs']) 
bleurt_extabs = average(scores)

# scores = scorer.score(references=details_dict['headlines'], candidates=details_dict['predict_ner']) 
# bleurt_ner = average(scores)

# scores = scorer.score(references=details_dict['headlines'], candidates=details_dict['predict_t5']) 
# bleurt_t5 = average(scores)

scores = scorer.score(references=details_dict['headlines'], candidates=details_dict['predict_art']) 
bleurt_art = average(scores)

bleurt_score = {'baseline (1s)': {'bleurt_score': bleurt_1s}, 
                'baseline (3s)': {'bleurt_score': bleurt_3s}, 
    'text' : {'bleurt_score': bleurt_text}, 
    'ext' : {'bleurt_score': bleurt_ext},
    'abs' : {'bleurt_score': bleurt_abs},
    'extabs' : {'bleurt_score': bleurt_extabs},
#    'ner' : {'bleurt_score': bleurt_ner},
#    't5' : {'bleurt_score': bleurt_t5},
    'art' : {'bleurt_score': bleurt_art},                
}

# bleurt_score = {'bleurt_score': {'baseline': bleurt_baseline}, 
#    'bleurt_score' : {'ext': bleurt_ext}, 
#   'bleurt_score' : {'abs': bleurt_abs},
#   'bleurt_score' : {'extabs': bleurt_extabs}
# }

#bleurt_score = {'bleurt_score': {'baseline': bleurt_baseline, 'ext': bleurt_ext, 'abs': bleurt_abs, 'extabs': bleurt_extabs}}

#df_results = {k: bleurt_score[k] for k in ('bleurt_score')}
df_results = {k: bleurt_score[k] for k in ('art', 'text', 'ext', 'baseline (3s)', 'abs', 'extabs', 'baseline (1s)')}
pd.DataFrame(df_results)

#score = nltk.translate.meteor_score.meteor_score(reference, model)
# print (nltk.translate.meteor_score.meteor_score(
#     ["this is an apple", "that is an apple"], "an apple on this tree"))
# print (nltk.translate.meteor_score.meteor_score(
#     ["this is an apple", "that is an apple"], "a red color fruit"))
# nltk.translate.meteor_score.meteor_score(P1, "a red color fruit")

# meteor_score_1s = 0
# for x, y in zip(details_dict['headlines'], details_dict['predict_1s']):
#     meteor_score_1s += nltk.translate.meteor_score.meteor_score(x, y)
# meteor_score_1s = meteor_score_1s/len(details_dict['headlines'])
# meteor_score_1s

# meteor_score_3s = 0
# for x, y in zip(details_dict['headlines'], details_dict['predict_3s']):
#     meteor_score_3s += nltk.translate.meteor_score.meteor_score(x, y)
# meteor_score_3s = meteor_score_3s/len(details_dict['headlines'])
# meteor_score_3s

# meteor_score_text = 0
# for x, y in zip(details_dict['headlines'], details_dict['predict_text']):
#     meteor_score_text += nltk.translate.meteor_score.meteor_score(x, y)
# meteor_score_text = meteor_score_text/len(details_dict['headlines'])
# meteor_score_text

# meteor_score_ext = 0
# for x, y in zip(details_dict['headlines'], details_dict['predict_ext']):
#     meteor_score_ext += nltk.translate.meteor_score.meteor_score(x, y)
# meteor_score_ext = meteor_score_ext/len(details_dict['headlines'])
# meteor_score_ext

# meteor_score_abs = 0
# for x, y in zip(details_dict['headlines'], details_dict['predict_abs']):
#     meteor_score_abs += nltk.translate.meteor_score.meteor_score(x, y)
# meteor_score_abs = meteor_score_abs/len(details_dict['headlines'])
# meteor_score_abs

# meteor_score_extabs = 0
# for x, y in zip(details_dict['headlines'], details_dict['predict_extabs']):
#     meteor_score_extabs += nltk.translate.meteor_score.meteor_score(x, y)
# meteor_score_extabs = meteor_score_extabs/len(details_dict['headlines'])
# meteor_score_extabs

# meteor_score_ner = 0
# for x, y in zip(details_dict['headlines'], details_dict['predict_ner']):
#     meteor_score_ner += nltk.translate.meteor_score.meteor_score(x, y)
# meteor_score_ner = meteor_score_ner/len(details_dict['headlines'])
# meteor_score_ner

# meteor_score_t5 = 0
# for x, y in zip(details_dict['headlines'], details_dict['predict_t5']):
#     meteor_score_t5 += nltk.translate.meteor_score.meteor_score(x, y)
# meteor_score_t5 = meteor_score_t5/len(details_dict['headlines'])
# meteor_score_t5

# meteor_score_art = 0
# for x, y in zip(details_dict['headlines'], details_dict['predict_art']):
#     meteor_score_art += nltk.translate.meteor_score.meteor_score(x, y)
# meteor_score_art = meteor_score_art/len(details_dict['headlines'])
# meteor_score_art

# meteor_score = {'baseline (1s)': {'meteor_score': meteor_score_1s}, 
#                 'baseline (3s)': {'meteor_score': meteor_score_3s}, 
#     'text' : {'meteor_score': meteor_score_text}, 
#     'ext' : {'meteor_score': meteor_score_ext},
#     'abs' : {'meteor_score': meteor_score_abs},
#     'extabs' : {'meteor_score': meteor_score_extabs},
#     'ner' : {'meteor_score': meteor_score_ner},
#     't5' : {'meteor_score': meteor_score_t5},
#     'art' : {'meteor_score': meteor_score_art}
# }
# df_results = {k: meteor_score[k] for k in ('baseline (1s)', 'baseline (3s)', 'text', 'ext', 'abs', 'extabs', 'ner', 't5', 'art')}
# pd.DataFrame(df_results)

len(details_dict['headlines'])
